color buff

+++ Assets/Resources/Scripts/Combat/SkillMasterShield.cs
@@ -27,7 +27,7 @@ public class SkillMasterShield : Skill
                 var shieldHp = (int)(owner.maxHp * skillCfg.Strength);
                 UnityEngine.Debug.Log("BattleBegin 护盾值 " + shieldHp);
 
-                BuffManager.AddBuff(unit, owner, (int)skillCfg.BuffId, skillCfg.BuffTime, shieldHp);
+                BuffManager.AddBuff(unit, owner, skillCfg.BuffId, skillCfg.BuffTime, shieldHp);
             }
 
         }
